LayoutTests/imported/w3c/web-platform-tests/dom/nodes/MutationObserver-document-expected.txt

MutationObservers: document mutations


PASS setup test
PASS parser insertion mutations
FAIL parser script insertion mutation assert_equals: target node must match expected Element node <p id="n00"><script id="inserted_script">document.body.ap... but got Element node <body><h1>MutationObservers: document mutations</h1>
<div...
PASS removal of parent during parsing
